Remove the 'util' package.
Distribute the functions and types in the 'util' package to the main package
and s3wrapper package. Also un-exports functions that are not used outside
those packages.
